+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Reworked API for dynamic objects.
   * Support for unmanaged dynamic objects through `DynamicMutableRealmObject.create()`.
   * Replaced `DynamicMutableRealm.create()` with `DynamicMutableRealm.copyToRealm()` similar to `MutableRealm.copyToRealm()`.
+* Moved `io.realm.MutableRealm.UpdatePolicy` to top-level class `io.realm.UpdatePolicy` as it now also applies to `DynamicMutableRealm.copyToRealm()`
 
 ### Enhancements
 * [Sync] Added support for `ObjectId` ([#652](https://github.com/realm/realm-kotlin/issues/652)). `ObjectId` can be used as a primary key in model definition.